SAURON-148:{SYSTEM}external interrupt survey[1/1]
[JIRA]
SAURON-148

[SUMMARY]
Monitoring changing of an external signal is important to the MCU system, then we need survey a method to do that. I chose a very simple example to show that.
At the start board, the user key is a suitable interrupt source. It's connected with MCU PA0. We configured the PA0 with EINT0, and added a handler for it.
Once the handler be triggered, a callback function will be activated. The callback will change the LED task's behaving -- its blink speed will be changed.

[Platform]
Saip

[Test]
local verify OK

[Feature]
SYSTEM

Change-Id: Ib0f2a3bafd5c6eb7d4eafe01382a4db68cb32871