a memory overflow issue fixed.
When the host sends continuous random data to the MCU to test its robustness, the MCU crashes. After inspection, the return value of the vsnprintf function is misunderstood by me. Re-test after correction, the issue disappears.

Because the stack usage of each thread is found to be close to the upper limit through tracking, the stack size of each thread is increased. Same thing for the UART sending/receiving message queue -- they changed with max size of datagram.

TODO: Sending should be done in an individual thread to avoid low-speed UART IO to reduce the caller's real-time performance.

Change-Id: If812644461b4ab6a7ce06e0a03f46d8709ea1c88