The protocol version query service and static version query mechanism are both OK.
for static version query mechanism, because SAIP needs to update the firmware on SACP, SACP needs to refer to the current firmware version and the firmware version that has been burned to decide whether to burn, so we need a simple way to let the SACP end know the firmware version number of the two, even if it is in When the firmware is not running yet. SACP can read the data of FLASH (that is, the location where the firmware is stored) through the ISP process, so our solution is to put the firmware version in the fixed position of the firmware to be burned, and the fixed position of the firmware that has been burned. It is feasible to put this position here in the startup.

The version information at a fixed location will also be used in the service of the protocol.

Change-Id: If4c07828733b46dc98fd060fbb696555fc8a77ee