Protocol running info collection and retrieving done.
run info is the counting data of some information in the data link layer (UART) and datagram layer during the operation of the protocol (including the host side and the MCU side). These count data are stored in an internal array. Those data of MCU can be accessed at the host by the msg_id GET_RUN_INFO.

Change-Id: I02595b10cbc97084ddcd8c27a1a28adc786f1c2e