SAURON-237:{EEPROM}changes EEPROM read interface adapted to factory test in PCBA[1/3]
[JIRA]
SAURON-237

[SUMMARY]
The production line personnel added a requirement: In PCBA production, the written PSN number needs to be read out for verification. To this end, a read interface has been added.
The EEPROM I2C address has been changed and also revised.
It was found that there was an error in the RCU initialization function before, and it was fixed.

[Platform]
Saip

[Test]
local verify OK

[Feature]
EEPROM

Change-Id: I59f7a03216ce9f2f7e72acc795431a0723292a91